package/libcamera: bump to 0.5.0
Release notes:
https://lists.libcamera.org/pipermail/libcamera-devel/2025-April/049672.html

There's now a pipeline for the Raspberry Pi 5 ISP (PiSP), that can be
added later on in a separate patch.

v4l2 compatibility layer is now a feature and not a boolean anymore. The
former takes "enabled", "disabled" or "auto"[1] while the latter takes
"true" or "false"[2] so we need to update how we configure the project to
match those expectations.

[1] https://mesonbuild.com/Build-options.html#features
[2] https://mesonbuild.com/Build-options.html#booleans

Signed-off-by: Quentin Schulz <quentin.schulz@cherry.de>
Signed-off-by: Thomas Petazzoni <thomas.petazzoni@bootlin.com>