configs/hifive_unmatched: add missing uboot custom hash file
Commit [1] introduced the hifive_unmatched_defconfig with
BR2_DOWNLOAD_FORCE_CHECK_HASH enabled and with custom hash files.
But it forgot to add the custom hash for uboot. It was not detected
at that time because the U-Boot version was the same as the default
"latest" version and was providing its hash.

This defconfig fails to build since the default "latest" was bumped
in commit [2].

This commit fixes the issue by adding the missing uboot hash.

Fixes:
https://gitlab.com/buildroot.org/buildroot/-/jobs/9677167053

[1] https://gitlab.com/buildroot.org/buildroot/-/commit/f9cc5dbd453c80790c1f2f7e9b19a85ded5979be
[2] https://gitlab.com/buildroot.org/buildroot/-/commit/c14ce5346d969e660cc788f2f4d8619b27c805a2

Signed-off-by: Julien Olivain <ju.o@free.fr>