board/raspberrypi/readme.txt: add note for bootcode.bin
The second-stage bootloader is contained in the onboard EEPROM[1] since
the Raspberry Pi 4 (i.e. for pi4, pi400, cm4, cm4s and pi5).

Therefore, the file bootcode.bin[2] is needed for the former Raspberry
Pi only (i.e. pi1, pi2, pi3, pi0, pi0w and pizero2w).

This adds a note precising the file bootcode.bin is present for the
former versions of the Raspberry Pi.

[1]: https://www.raspberrypi.com/documentation/computers/raspberry-pi.html#raspberry-pi-boot-eeprom
[2]: https://www.raspberrypi.com/documentation/computers/configuration.html#bootcode-bin

Signed-off-by: Gaël PORTAY <gael.portay+rtone@gmail.com>
Signed-off-by: Julien Olivain <ju.o@free.fr>