configs/warp7: Improve Wifi instructions
The "modprobe brcmfmac" command is not necessary because this driver
is automatically loaded.

Remove the "iwconfig" line as it is considered deprecated:

warning: `iwconfig' uses wireless extensions which will stop working for Wi-Fi 7 hardware; use nl80211

Remove the "-Dwext" parameter as it is not supported by the default kernel
configuration:

wlan0: Unsupported driver 'wext'

Tested Wifi by following the updated commands.

Signed-off-by: Fabio Estevam <festevam@gmail.com>
Signed-off-by: Julien Olivain <ju.o@free.fr>