package/dust: bump to version 1.1.2
Release:
https://github.com/bootandy/dust/releases/tag/v1.1.2

Note: version 0.9.0 of dust fails to build when running
the runtime test on the armv7 architecture due to an
unconditional import of Atomicu64.

	error:

Compiling config-file v0.2.3
error[E0432]: unresolved import `std::sync::atomic::AtomicU64`
 --> src/progress.rs:6:18
  |
6 |         atomic::{AtomicU64, AtomicU8, AtomicUsize, Ordering},
  |                  ^^^^^^^^^
  |                  |
  |                  no `AtomicU64` in `sync::atomic`
  |                  help: a similar name exists in the module: `AtomicU32`

For more information about this error, try `rustc --explain E0432`.
error: could not compile `du-dust` (bin "dust") due to 1 previous error

This issue was discovered while writing a runtime test
for dust. upgrading to version 1.1.2 resolves the issue.

More details available in the following issue:
https://github.com/bootandy/dust/issues/423

For now, we bump to the latest compatible version
which builds and runs correctly. We can't bump to the latest
version 1.2.0 since it requires a cargo version newer than
1.82.0.

	error:
--  The package requires the Cargo feature called `edition2024`, but that feature is not stabilized in this version of Cargo (1.82.0 (8f40fc59f 2024-08-21)).
  Consider trying a newer version of Cargo (this may require the nightly release).

The upgrade to 1.2.0 will be considered once the patch for
Rust 1.86.0 is accepted.

Signed-off-by: El Mehdi YOUNES <elmehdi.younes@smile.fr>
Signed-off-by: Julien Olivain <ju.o@free.fr>