package/libqrencode: fix the _SITE package download url
The qrencode release archive is no longer distributed at [1].
This url now shows a HTTP/404 not found error.

The qrencode homepage [2] states:
"You can download the source packages at Releases[3] page at GitHub."

This commit updates the package download url to use github. The github
archive is not the same as the old release archive: it does not
contain the autotools configure script. For this reason, the hash
archive is updated and "_AUTORECONF = YES" is also added.

This commit also fixes the zbar runtime test [4] which is using
libqrencode to generate a qrcode to test zbar.

Fixes: [4]

[1] http://fukuchi.org/works/qrencode/qrencode-4.1.1.tar.gz
[2] https://fukuchi.org/en/works/qrencode/
[3] https://github.com/fukuchi/libqrencode/releases
[4] https://gitlab.com/buildroot.org/buildroot/-/jobs/9774022888

Signed-off-by: Julien Olivain <ju.o@free.fr>
Signed-off-by: Thomas Petazzoni <thomas.petazzoni@bootlin.com>