package/heimdal: bump version to 8c3c97bdf6c06200418f1a85aa22beaa441c6b23
The current version was more than 3 years old and had known CVEs (see
CVE-2023-36328). Since this is a host only package, we don't consider
this as a security bump.

This upgrade allows the removal of the two patches currently applied:

- The commit 1b57b62, which fixes the build with autoconf 2.72, is
  already included upstream.

- We no longer need to use the JSON-PP module, as it is now
  optional (commit 13d3bcf).

Also, since the latest version of Heimdal no longer depends on
e2fsprogs, the host-e2fsprogs dependency has been removed.

Signed-off-by: Guillaume Chaye <guillaume.chaye@zeetim.com>
Signed-off-by: Thomas Petazzoni <thomas.petazzoni@bootlin.com>