rpi-firmware: support installing extra files
There are several use cases for installing additional files in the boot
partition that is read by the RPi firmware.

- autoboot.txt is an optional configuration file for the RPi
  firmware [1]. Supporting several autoboot files will enable A/B
  setups, as using the renameat2() system call with the RENAME_EXCHANGE
  flag will let users atomically replace one autoboot configuration
  file with the other. This improves reliability in the case of an
  update which could potentially be interrupted.

- Multiple cmdline.txt files are useful in the context of a new
  [boot_partition] conditional filter introduced in config.txt in
  commit [2]. This is useful for A/B systems to have identical BootFS
  partitions on both slots, and not have to edit the kernel command line
  to ensure the kernel will load the right rootFS after update of the
  BootFS.

- rpi-firmware contains DTB overlays for many "standard" hats, but a
  custom hat may require a custom overlay.

Although it is possible to install additional files in the boot
partition in the post-image script, it is very convenient to be able to
use the standard RPi post-image script in
board/raspberrypi/post-image.sh. That script looks in
$BINARIES_DIR/rpi-firmware, so it is convenient to be able to place
additional files there.

Add the option BR2_PACKAGE_RPI_FIRMWARE_EXTRA_FILES which is simply a
list of files to be copied to $BINARIES_DIR/rpi-firmware, which will
eventually end up as the boot partition. Make sure that this is done as
the last step of RPI_FIRMWARE_INSTALL_IMAGES_CMDS, so the files can
override files installed by earlier steps.

[1] https://www.raspberrypi.com/documentation/computers/config_txt.html#autoboot-txt
[2] https://github.com/raspberrypi/rpi-eeprom/commit/d50b2b32f162292ab6e235932075ce00bddda4ae

Signed-off-by: Olivier Benjamin <olivier.benjamin@bootlin.com>
Signed-off-by: Arnout Vandecappelle <arnout@rnout.be>